Move circles to component
