+++ Application/src/Screens/ConnexionScreen.tsx
@@ -7,6 +7,7 @@ import GoogleSocialButton from "../Components/SocialButtons/GoogleSocialButton";
 import AppleSocialButton from "../Components/SocialButtons/AppleSocialButton";
 import { NavigatorPush } from "../Navigator";
 import { Options } from "react-native-navigation";
+import Circles from "../Components/Circles";
 
 
 export default function ConnexionScreen() {
@@ -68,15 +69,6 @@ export default function ConnexionScreen() {
                         <Text style={signStyles.mobileSignInText}>S'inscrire</Text>
                     </TouchableOpacity>
                 </View>
-            </View>
-        )
-    }
-
-    function Circles() {
-        return (
-            <View>
-                <View style={circles.bigCircle}/>
-                <View style={circles.smallCircle}/>
             </View>
         )
     }
@@ -226,28 +218,5 @@ const signStyles = StyleSheet.create({
         color: "#95B8D1",
         fontFamily: "Poppins-Bold",
         fontSize: 18
-    }
-})
-
-const circles = StyleSheet.create({
-    bigCircle: {
-        width: 406,
-        height: 406,
-        borderRadius: 500,
-        backgroundColor: Globals.Colors.main,
-        position: "absolute",
-        top: -244,
-        left: -55,
-        opacity: 0.4
-    },
-    smallCircle: {
-        width: 342,
-        height: 342,
-        borderRadius: 500,
-        backgroundColor: Globals.Colors.main,
-        position: "absolute",
-        top: -203,
-        left: 179,
-        opacity: 0.4
     }
 })
